web/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
geoalchemy2==0.14.1
geopandas==0.14.1 